requirements.txt
ipython
numpy
scipy
matplotlib
cv2
solidpython